+++ app/Main.hs
@@ -20,12 +20,15 @@ data DumpTarget = Linker { _pkgGet :: String }
                 | LdLibPath { _pkgGets :: [String] }
 
 data Command = Install { _pkgName :: String, _verbosity :: Verbosity, _target :: Maybe Platform, _static :: Bool, _global :: Bool, _packageSet :: Maybe String }
+             | Uninstall { _pkgStr :: String, _verbosity :: Verbosity, _target :: Maybe Platform }
              | Check { _dhallFile :: String, _verbosity :: Verbosity }
              | CheckSet { _dhallFile :: String, _verbosity :: Verbosity }
              | Dump { _dumpTarget :: DumpTarget, _host :: Maybe Platform }
              | DumpCabal { _pkgGetsCabal :: [String], _host :: Maybe Platform }
              | List { _packageSet :: Maybe String }
              | Nuke
+             | NukeCache
+             | GarbageCollect { _verbosity :: Verbosity }
 
 verbosityInt :: Parser Int
 verbosityInt = length <$>
@@ -63,12 +66,15 @@ dumpTarget = hsubparser
 userCmd :: Parser Command
 userCmd = hsubparser
     (command "install" (info install (progDesc "Install a package from the global package set"))
+    <> command "uninstall" (info uninstall (progDesc "Uninstall a package"))
     <> command "check" (info check (progDesc "Check a Dhall expression to ensure it can be used to build a package"))
     <> command "check-set" (info checkSet (progDesc "Check a package set defined in Dhall"))
     <> command "dump" (info dump (progDesc "Display flags to link against a particular library"))
     <> command "dump-cabal" (info dumpCabal (progDesc "Display flags to use with cabal new-build"))
     <> command "list" (info list (progDesc "List all available packages"))
     <> command "nuke" (info (pure Nuke) (progDesc "Remove all globally installed libraries"))
+    <> command "nuke-cache" (info (pure NukeCache) (progDesc "Remove cached soure tarballs"))
+    <> command "garbage-collect" (info garbageCollect' (progDesc "Garbage collect redundant packages"))
     )
 
 list :: Parser Command
@@ -79,6 +85,15 @@ ftypeCompletions ext = completer . bashCompleter $ "file -X '!*." ++ ext ++ "' -
 
 dhallCompletions :: Mod ArgumentFields a
 dhallCompletions = ftypeCompletions "dhall"
+
+uninstall :: Parser Command
+uninstall = Uninstall
+    <$> argument str
+        (metavar "PACKAGE"
+        <> help "Name of package to uninstall"
+        <> completer (listIOCompleter allPackages))
+    <*> verbosity
+    <*> target
 
 install :: Parser Command
 install = Install
@@ -107,6 +122,9 @@ static' =
     switch
     (long "static"
     <> help "Build static libaries")
+
+garbageCollect' :: Parser Command
+garbageCollect' = GarbageCollect <$> verbosity
 
 check :: Parser Command
 check = Check <$> dhallFile <*> verbosity
@@ -149,6 +167,9 @@ dhallFile =
     )
 
 run :: Command -> IO ()
+run (Uninstall pkId v host') = do
+    parsedHost <- parseHostIO host'
+    runPkgM v $ uninstallPkgByName pkId parsedHost
 run (Install pkId v host' sta glob pkSet) = do
     parsedHost <- parseHostIO host'
     runPkgM v $ buildByName (T.pack pkId) parsedHost pkSet sta glob
@@ -166,7 +187,9 @@ run Nuke = do
     exists <- doesDirectoryExist pkgDir
     when exists $
         removeDirectoryRecursive pkgDir
+run NukeCache = cleanCache
 run (List pkSet) = displayPackageSet pkSet
+run (GarbageCollect v) = runPkgM v garbageCollect
 
 main :: IO ()
 main = run =<< execParser wrapper