+++ src/Package/C.hs
@@ -29,6 +29,9 @@ module Package.C (
                  , printLdLibPath
                  , printCabalFlags
                  , buildByName
+                 , uninstallPkgByName
+                 , garbageCollect
+                 , cleanCache
                  -- * Dhall functionality
                  , cPkgDhallToCPkg
                  , getCPkg
@@ -44,6 +47,7 @@ module Package.C (
 
 import           Package.C.Build
 import           Package.C.Build.Tree
+import           Package.C.Db.GarbageCollect
 import           Package.C.Db.Monad
 import           Package.C.Db.Register
 import           Package.C.Db.Type