+++ src/Package/C/Db/Type.hs
@@ -32,4 +32,6 @@ data BuildCfg = BuildCfg { buildName       :: String
                          , configureCmds   :: [ Command ]
                          , buildCmds       :: [ Command ]
                          , installCmds     :: [ Command ]
+                         -- , tarball         :: FilePath
+                         , manual          :: Bool -- ^ Was this package manually installed?
                          } deriving (Eq, Ord, Generic, Binary, Hashable)