+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # cpkg
 
+## 0.2.3.0
+
+  * Add `garbage-collect` subcommand
+  * Add `uninstall` subcommand
+  * Add `nuke-cache` subcommand
+
 ## 0.2.2.0
 
   * Add `printLdLibFlags` function and add functionality to CLI interface